refactor: 검증이 Baseball에 종속적이므로 GameValidator -> BaseballGuide 수정
- BaseballGuide의 Inner Class로 Validator가 존재
- Baseball 상수를 BaseballGuide에서 관리
